refactor(web): スケジュール描画を共通化し、共有画面のフォーク複製を解消
Figma デザイン反映 (第4段) の下準備。見た目は変えず、構造だけを整える。
デザインそのものの反映は後続 PR で行う。

## 何が問題だったか

スケジュールのボード描画が 2 箇所にあった。

- ItemSchedulePage.tsx (1279行) … 認証済み画面
- ShareSchedule.tsx (487行)    … 共有リンク画面。上のフォーク複製

同じレイアウトを二重に持っているため、デザイン変更のたびに同じ修正を 2 回
行う必要があり、実際に両者は既にずれていた。

## やったこと

ボード描画を features/plans/schedule/ に切り出し、両画面で共有する。
ScheduleBoard に editing を渡せば編集モード、渡さなければ閲覧専用になる。

  ScheduleBoard  … ボード本体 (ドラッグ移動/期間リサイズ/後続紐づけ/チェーン強調)
  DateAxis       … 縦軸
  ColumnHeader   … 制作物列ヘッダー
  BallChip       … ボール。mode='edit' | 'view'
  LinkLayer      … 後続コネクトの SVG
  ZoomControl    … 行高スライダー
  chain.ts       … チェーン探索・紐づけ可否判定・保持者解決 (純粋関数)
  dayTones.ts    … 週末/祝日/本日の判定

ItemSchedulePage 1279行 → 388行、ShareSchedule 487行 → 69行。

## 統合で共有画面の挙動が揃った点 (いずれも認証済み画面に合わせた)

1. 列ヘッダーのボール保持者
   旧: 最後の active な予定の ballHolder をそのまま表示
   新: deriveLineBallHolders でライン単位に導出 (#117 のモデル)
   実データでは server の ballHolder も同じ導出なので表示は変わらないが、
   後続チェーンが複数ラインに分かれる場合は新しい方が正しい。

2. カードの縦位置
   旧: top = idx*rowHeight + 1 / height = ... - 3 (独自の数値)
   新: chipVerticalBounds (CARD_INSET_Y=4) に統一
   後続リンクの線とカードのわずかなずれが直る。

3. カード下部の 2 行目
   旧: 実施者 / 保持者
   新: 実施者 / 承認者 (無ければ進行責任者) + 保持者バッジ
   共有 API は元から同じ PlanDTO (approver / progressManager 込み) を返して
   いるため、新たな情報開示にはならない。

## ついでに直したもの

後続矢印の SVG marker を列ごとの <defs> に置いていたため、列の数だけ同じ
DOM id が重複していた (旧 ShareSchedule はこれを避けるため別 id を使って
いた)。ボード直下に 1 つだけ置く形に集約した。

## 確認

- lint / type-check / test (676 tests) / カバレッジ閾値 通過
- ShareSchedule のテストは上記 1〜3 の挙動統一に合わせて更新した
